Simplify logic and allow any SPDX header by default
As a user, I found the inclusion of _author_ confusing, and I had to check the code to see how it gets included. I think it might be simpler if the regex is the only parameter. There is a common recommendation to just list the authors as "The Authors" or "Contributors to <Project>" anyway, both for legal reasons (sometimes who actually holds the copyright would be decided in court), because lots more people than the authors could hold part of the copyright, because you'd have to update the header if the authors change, and because it's not necessary.

While I was at it, I also set the default min length to 1024, closed the open file handle, dropped the unnecessary encoding header, and included any `SPDX-` header in the default regex.